Create data analysis and visualization project
This coursework aims to prepare data for analysis, implement, a model and tell a story by visualising the data within the chosen scenario. 
Scenario: A topic needs to be chosen by the student and agreed with the workshop tutor; students are allowed to pick any topic they are interested in, and they are required to find their own data from any source they want to. 

Some example data sources are:
	https://www.data.gov.uk/
	https://data.london.gov.uk/
	https://www.ons.gov.uk/
	https://www.kaggle.com/


Some example topics are: 
	Grouping customers by similarity to avoid customer retention
	Finding items to sell together at a local clothing shop
	What is the profit trend of a particular product?
	Is there a correlation between life expectancy and education level?


Data analysis requirements and marking scheme:
1.	Academic Journal : 
a.	At the end of the analysis, students are required to write an academic journal in IEEE referencing style. The journal should academically tell the story.
b.	Journal should include an introduction, the model(s) used – literature review and methodology, the findings and critical reflection with recommendations.

2.	Presentation : Students will also tell their story with a presentation to their workshop group, presentation should be no longer than 10 minutes. All students in the audience will be expected to ask questions and give feedback to the presenter.